+++ R/Aaaa.R
@@ -1,10 +1,11 @@
 ##' @importFrom grDevices pdf dev.off graphics.off postscript
-##' @importFrom ggplot2 theme margin ggplotGrob
+##' @importFrom ggplot2 theme margin ggplotGrob last_plot
 ##' @importFrom gridExtra grid.arrange arrangeGrob marrangeGrob
 ##' @importFrom grid grid.draw gpar grid.newpage
 ##' @importFrom grid textGrob gList
 ##' @importFrom assertthat assert_that
 ##' @importFrom rlang flatten flatten_if
+##' @importFrom glue glue
 ##'
 NULL